release: move DMG to CI, simplify local install
- publish.py → install.py (local install only)
- DMG build/upload moves to release.yml (new build-dmg job)
- Screenshots command moves to dev.py
- DMG attached to GitHub releases alongside native binaries
- R2 upload uses region_name='auto' (fixes Cloudflare R2 compatibility)
